Update compiler diagnostic flags and fix type inconsistencies
This change updates the compiler diagnostic flags in the pch.h file to ignore warnings related to switch statements and function with variable arguments. It also corrects a type inconsistency by casting the loop variable in ByteCodeRun_FFI.cpp to uint32_t and properly initializing the va_list in ErrorIds.cpp.

+++ src/ErrorIds.cpp
@@ -1091,7 +1091,7 @@ void initErrors()
 Utf8 formErr(ErrorID idx, ...)
 {
     va_list args;
-    va_start(args, static_cast<uint64_t>(idx));
+    va_start(args, idx);
     auto result = form("[%s] %s", g_EI[idx], g_E[idx].c_str());
     result      = form(result.c_str(), args);
     va_end(args);
